Udacity Android Dev ND Sandwich Club Project
Project Overview:

An app which shows a list of sandwich name and shows the details of each sandwich once it is selected.

Steps Taken:

Designed the layout for the detail activity so the different elements are displayed in a sensible way. 
Implemented the JSON parsing in JsonUtils so it produces a Sandwich Object that can be used to populate the UI that was designed.
Created a landscape xml file for the detail activity so that the activity looks nice when the phone is held in landscape position.
Added a ScrollView to the activity_detail.xml file to ensure that all the details are visible on smaller devices.

Disclaimer and Warnings:

This project was done as part of Udacity's Android Developer NanoDegree. Starter code was provided to me by Udacity for this project. My additions were the JsonUtils.java, the DetailActivity.java and activity_detail.xml files, the styles.xml, the dimens.xml files, and some of the strings.xml file. I also changed the colour scheme of the app.